% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr_plot.R
\name{autoplot.ttt_board}
\alias{autoplot.ttt_board}
\title{Autoplot method for a 4D tic-tac-toe board}
\usage{
\method{autoplot}{ttt_board}(object, ...)
}
\arguments{
\item{object}{A \code{ttt_board}.}

\item{...}{Passed to \code{tsr_plot()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Registers a \code{ggplot2::autoplot()} method on \code{ttt_board} that delegates to
\code{tsr_plot()}.
}
\examples{
ggplot2::autoplot(tsr_new_board())
}
