% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr_board.R
\name{is_ttt_board}
\alias{is_ttt_board}
\title{Test whether an object is a \code{ttt_board}}
\usage{
is_ttt_board(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar: \code{TRUE} if \code{x} inherits from \code{"ttt_board"}.
}
\description{
Test whether an object is a \code{ttt_board}
}
\examples{
is_ttt_board(tsr_new_board())
is_ttt_board(list())
}
