% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr_status.R
\name{tsr_is_full}
\alias{tsr_is_full}
\title{Whether the board has no empty cells}
\usage{
tsr_is_full(board)
}
\arguments{
\item{board}{A \code{ttt_board}.}
}
\value{
Logical scalar.
}
\description{
Whether the board has no empty cells
}
\examples{
tsr_is_full(tsr_new_board())
}
